refactor(api): enhance data access and support service structure
